+++ prisma/patches/index.js
@@ -1,0 +1,2 @@
+await import("./item-price.js");
+await import("./list-relationship.js");
